Refactoring changes WIP

+++ scripts/bundle_reader.sh
@@ -6,13 +6,27 @@ realpath() {
 
 SCRIPT_PATH=`realpath "$0"`
 SCRIPT_DIR=`dirname "$SCRIPT_PATH"`
+READER_SUBMODULE_DIR="$SCRIPT_DIR/../reader"
 READER_DIR="$SCRIPT_DIR/../bundled/reader"
+HASH_FILE="$READER_DIR/reader_hash.txt"
+CURRENT_HASH=`git ls-tree --object-only HEAD "$READER_SUBMODULE_DIR"`
 
 if [ -d "$READER_DIR" ]; then
- rm -rf "$READER_DIR"
+    if [ -d "$HASH_FILE" ]; then
+        CACHED_HASH=`cat "$HASH_FILE"`
+    else
+        CACHED_HASH=0
+    fi
+
+    if [ $CACHED_HASH == $CURRENT_HASH ]; then
+        exit
+    else
+        rm -rf "$READER_DIR"
+    fi
 fi
 
-cd "$SCRIPT_DIR/../reader"
+cd "$READER_SUBMODULE_DIR"
 NODE_OPTIONS=--openssl-legacy-provider npm ci
 NODE_OPTIONS=--openssl-legacy-provider npm run build:ios
-mv "$SCRIPT_DIR/../reader/build/ios" "$READER_DIR"
+mv "$READER_SUBMODULE_DIR/build/ios" "$READER_DIR"
+echo "$CURRENT_HASH" > "$HASH_FILE"
